# [WIP] Mixed-precision radiation module