--- a/include/picongpu/plugins/radiation/Radiation.kernel
+++ b/include/picongpu/plugins/radiation/Radiation.kernel
@@ -468,14 +468,14 @@ namespace radiation
 
                                 // complex amplitude for j-th particle
                                 // LOCAL can be single precision
-                                Amplitude<picongpu::float_32> amplitude_add(
+                                Amplitude<> amplitude_add(
                                     weighted_real_amp,
                                     t_ret_s[ j ] * omega
                                 );
 
                                 // add this single amplitude those previously considered
-                                amplitude += precisionCast<picongpu::float_64>(amplitude_add);
-                                //amplitude += amplitude_add;
+                                //amplitude += precisionCast<picongpu::float_64>(amplitude_add);
+                                amplitude += amplitude_add;
 
                             }// END: check Nyquist-limit for each particle "j" and each frequency "omega"