--- a/include/picongpu/plugins/radiation/VectorTypes.hpp
+++ b/include/picongpu/plugins/radiation/VectorTypes.hpp
@@ -33,6 +33,7 @@ namespace radiation
     using vector_X = cuda_vec< picongpu::float3_X, picongpu::float_X >;
     using vector_32 = /*__align__(16)*/ cuda_vec< picongpu::float3_32, picongpu::float_32 >;
     using vector_64 = /*__align__(32)*/ cuda_vec< picongpu::float3_64, picongpu::float_64 >;
+    using vector_32_64 = cuda_vec< picongpu::float3_32, picongpu::float_64 >;
 } // namespace radiation
 } // namespace plugins
 } // namespace picongpu