--- a/include/picongpu/plugins/radiation/amplitude.hpp
+++ b/include/picongpu/plugins/radiation/amplitude.hpp
@@ -34,13 +34,15 @@ namespace radiation
 
 /** class to store 3 complex numbers for the radiated amplitude
  */
+template<typename base_dtype = picongpu::float_64 >
 class Amplitude
 {
 public:
-  using complex_64 = pmacc::math::Complex< picongpu::float_64 >;
-
+  // For the intermediate amplitude values we may use single precision
+  // for the final accumulation we will have to use double precision
+  using complex_T = pmacc::math::Complex< base_dtype >;
   /* number of scalar components in Amplitude = 3 (3D) * 2 (complex) = 6 */
-  static constexpr uint32_t numComponents = uint32_t(3) * uint32_t(sizeof(complex_64) / sizeof(typename complex_64::type));
+  static constexpr uint32_t numComponents = uint32_t(3) * uint32_t(sizeof(complex_T) / sizeof(typename complex_T::type));
 
   /** constructor
    *
@@ -52,11 +54,10 @@ public:
       picongpu::float_X cosValue;
       picongpu::float_X sinValue;
       pmacc::math::sincos(phase, sinValue, cosValue);
-      amp_x=pmacc::math::euler(vec.x(), precisionCast<picongpu::float_64>(sinValue), precisionCast<picongpu::float_64>(cosValue) );
-      amp_y=pmacc::math::euler(vec.y(), precisionCast<picongpu::float_64>(sinValue), precisionCast<picongpu::float_64>(cosValue) );
-      amp_z=pmacc::math::euler(vec.z(), precisionCast<picongpu::float_64>(sinValue), precisionCast<picongpu::float_64>(cosValue) );
+      amp_x=pmacc::math::euler(precisionCast<base_dtype>(vec.x()), precisionCast<base_dtype>(sinValue), precisionCast< base_dtype >(cosValue) );
+      amp_y=pmacc::math::euler(precisionCast<base_dtype>(vec.y()), precisionCast<base_dtype>(sinValue), precisionCast<base_dtype>(cosValue) );
+      amp_z=pmacc::math::euler(precisionCast<base_dtype>(vec.z()), precisionCast<base_dtype>(sinValue), precisionCast<base_dtype>(cosValue) );
   }
-
 
   /** default constructor
    *
@@ -79,6 +80,16 @@ public:
 
   }
 
+  /** constructor
+   *
+   *  @param x
+   *  @param y
+   *  @param z
+   */
+  HDINLINE Amplitude(const complex_T& x,
+                     const complex_T& y,
+                     const complex_T& z)
+          :amp_x(x),amp_y(y),amp_z(z){ }
 
   /** returns a zero amplitude vector
    *
@@ -86,9 +97,9 @@ public:
   HDINLINE static Amplitude zero(void)
   {
       Amplitude result;
-      result.amp_x = complex_64::zero();
-      result.amp_y = complex_64::zero();
-      result.amp_z = complex_64::zero();
+      result.amp_x = complex_T::zero();
+      result.amp_y = complex_T::zero();
+      result.amp_z = complex_T::zero();
       return result;
   }
 
@@ -133,11 +144,22 @@ public:
       return amp_x.get_real();
   }
 
+  /* Getters for the components
+   */
+  HDINLINE complex_T getXcomponent() const{
+          return this->amp_x;
+  }
+  HDINLINE complex_T getYcomponent() const{
+          return this->amp_y;
+  }
+  HDINLINE complex_T getZcomponent() const{
+          return this->amp_z;
+  }
 
 private:
-  complex_64 amp_x; // complex amplitude x-component
-  complex_64 amp_y; // complex amplitude y-component
-  complex_64 amp_z; // complex amplitude z-component
+  complex_T amp_x; // complex amplitude x-component
+  complex_T amp_y; // complex amplitude y-component
+  complex_T amp_z; // complex amplitude z-component
 
 };
 } // namespace radiation
@@ -151,12 +173,61 @@ namespace mpi
 
   /** implementation of MPI transaction on Amplitude class */
   template<>
-  HINLINE MPI_StructAsArray getMPI_StructAsArray< picongpu::plugins::radiation::Amplitude >()
+  HINLINE MPI_StructAsArray getMPI_StructAsArray< picongpu::plugins::radiation::Amplitude<> >()
   {
-      MPI_StructAsArray result = getMPI_StructAsArray< picongpu::plugins::radiation::Amplitude::complex_64::type > ();
-      result.sizeMultiplier *= picongpu::plugins::radiation::Amplitude::numComponents;
+      MPI_StructAsArray result = getMPI_StructAsArray< picongpu::plugins::radiation::Amplitude<>::complex_T::type > ();
+      result.sizeMultiplier *= picongpu::plugins::radiation::Amplitude<>::numComponents;
       return result;
   };
 
 } // namespace mpi
+} // namespace pmacc
+
+
+
+namespace pmacc
+{
+namespace algorithms
+{
+namespace precisionCast
+{
+/* We want to be able to cast a low
+ * precision amplitude to a high precision one
+ */
+
+        template<typename CastToType>
+                struct TypeCast<
+                        CastToType,
+                        picongpu::plugins::radiation::Amplitude<CastToType>
+                        >
+                        {
+                                using result = const picongpu::plugins::radiation::Amplitude<CastToType>&;
+
+                                HDINLINE result operator( )( result amplitude ) const
+                                {
+                                        return amplitude;
+                                }
+                        };
+
+        template<typename CastToType,
+                typename OldType >
+                        struct TypeCast<
+                        CastToType,
+                        picongpu::plugins::radiation::Amplitude<OldType>
+                        >
+                        {
+                                using result = picongpu::plugins::radiation::Amplitude<CastToType>;
+                                using ParamType = picongpu::plugins::radiation::Amplitude<OldType>;
+                                HDINLINE result operator( )(const ParamType& amplitude ) const
+                                {       
+                                        result Result( 
+                                                    precisionCast<result::complex_T::type >(amplitude.getXcomponent()),
+                                                    precisionCast< result::complex_T::type>(amplitude.getYcomponent()),
+                                                    precisionCast< result::complex_T::type>(amplitude.getZcomponent()) );
+                                        return Result;
+                                }
+                        };
+
+} // namespace typecast
+} // namespace algorithms
 } // namespace pmacc