… include/picongpu/plugins/radiation/vector.hpp