## Intent
### Current problem
Quantities such as position `r` and observer direction `n` are up-cast by default to 64 bit precision.
The detector direction `n` is computed using 32 bit precision in `radiationObserver.param` but converted to `picongpu::float_64` vector at the end.

### Proposal
The intent of the changes is to introduce mixed (single/double) precision computations into the radiation module of PIConGPU.

Ideally the usage of only _sufficient_ precision for each separate computation will reduce register and memory consumption and allow more thread blocks to run in parallel.

Formal analysis suggests that most of the quantities can be kept in 32 bit precision (`float_X`) and at best only the accumulators of inner products and the results have to remain in 64 bit precision.

## Implementation

The basic data type of the `Amplitude` class is templated such that it can be switched as desired.
Arguments to the functions that compute the separate parts of the amplitude are adapted to take positions and observation direction of 32 bit precision.

The only quantity that has to be stored in double precision is the retarded time `t_ret_s`, even if it is computed from input quantities of lower precision.

**Changing `r` to a `vector_32` type in `Retarded_time_1` *will* inctroduce large errors.**

Usage of different precision for accumulated amplitude values, frequencies and observation directions requires different `radSplashType` definitions in `writeHDF5file`, lest the values are computed properly, but stored with a type-mismatch (which will lead to confusing outputs).

## Tests
### Correctness

A first check of corretness was performed using the single-particle configuration of the `Bunch` example of PIConGPU.
It was run for 7000 time steps, with radiation enabled from 2800 to 3000.

The quality of the computed amplitudes was tested by computing (for arbitrarily chosen steps 2900 and 3000) the maximum relative errors of the complex amplitudes in 2 and inf norms

<img src="https://render.githubusercontent.com/render/math?math=\max_{\vec{n},\omega} \left( \Vert \vec{A}_{new} - \vec{A}_{ref} \Vert_{2,\infty} / \Vert \vec{A}_{ref} \Vert_{2,\infty} \right)">

as well as maximum deviation w.r.t. the maximum achieved norm of the reference amplitude:

<img src="https://render.githubusercontent.com/render/math?math=\max_{\vec{n},\omega} \left( \Vert \vec{A}_{new} - \vec{A}_{ref} \Vert_{2,\infty} \right) / \max_{\vec{n},\omega} \left( \Vert \vec{A}_{ref} \Vert_{2,\infty} \right)">

This latter test does not emphasize deviations from reference solution in the regions of the spectrum where the intensity is negligible anyway.

The following errors were observed for the commit 50c0f50a at time step 3000:
```
    |A_tgt - A_ref|_2  3.8007873728778394e-22
    |A_tgt - A_ref|_inf  3.799745971153532e-22
    |A_ref|_2  2.507966173029181e-17
    |A_ref|_inf  2.507966173029181e-17
    |A_tgt|_2  2.5079660813987757e-17
    |A_tgt|_inf  2.5079660813987754e-17
    |A_tgt - A_ref|_2 / |A_ref|_2 0.006734829014505873
    |A_tgt - A_ref|_inf / |A_ref|_inf 0.006737196156335815
    max(|A_tgt - A_ref|_2) / max(|A_ref|_2) 1.5154858999900937e-05
    max(|A_tgt - A_ref|_inf) / max(|A_ref|_inf) 1.5150706624420332e-05
```

Hence the proper relative error is <img src="https://render.githubusercontent.com/render/math?math=6.73\cdot 10^{-3}">, whereas the error in units of maximum amplitude is <img src="https://render.githubusercontent.com/render/math?math=1.52\cdot 10^{-5}">.

### Performance

Performance of the implementation was tested using a thermal plasma (E = 0.511 [keV]) of density `2e23` with a grid of `1024 x 384 x 1024` cells distributed to 4 GPUs of _one_ V100 node according (`2 1 2` distribution) and 2 particles per cell. The simulation runs for 27 time steps with radiation enabled from the start. Only step 0 is written to disk.

The reference code required `1h 41min 48sec 233msec = 6108 sec` to complete the calculation.
The proposed changes reduce this to `1h 21min 10sec 663msec = 4870 sec` calculation time. 

A reduction of run time of ~20% and a **speed-up** of `~1.25x`.

This comes with **two observed caveats**:

1. Increasing the number of jobs for the reduction in the radiation kernel `--e_radiation.numJobs 4` will
result in the mixed-precision version taking longer to complete the same calculation as the reference version.
REF (4 jobs): ` 55 % =       15 | time elapsed:      40min 38sec 168msec | avg time per step:  2min 52sec 536msec`
MPRAD (4 jobs): ` 55 % =       15 | time elapsed:      42min 17sec 728msec | avg time per step:  2min 59sec 703msec`

2. Usage of a 32 bit `amplitude_add` in `Radiation.kernel` while working will slow the computation down
significantly:
REF (2 jobs): `55 % =       15 | time elapsed:      52min 31sec 526msec | avg time per step:  3min 47sec 368msec`
MPRAD (2 jobs, 64 bit increment): `55 % =       15 | time elapsed:      42min 16sec 466msec | avg time per step:  2min 59sec 564msec`
MPRAD (2 jobs, 32 bit increment): `55 % =       15 | time elapsed:   1h  5min 57sec 342msec | avg time per step:  4min 49sec 640msec`
This is likely due to the need to cast an `Amplitude` object from 32 bit to 64 bit immediately after it is created, using `precisionCast`. The latter produces a temporary `Amplitude` object with 64 bit precision of the components and constitutes an overhead.