@Anton-Le 2021-02-11
@PrometheusPi I would be grateful for input regarding further testing.
@psychocoderHPC Is it possible to parametrize the data type (here `float_64`) of the `typename PICToSplash<float_64>::type radSplashType;` in `Radiation.hpp` to always be determined by the type of of the data stored in `detectorPositions` and `detectorFrequencies` ? Ideally without templating the entire class.
Currently I just introduced 3 different `typename` statements, which does not benefit maintainability.