@Anton-Le 2021-02-11
# Further test for correctness

The version with 32 bit `amplitude_add` was also tested on an example of a tilted, underdense (n_b = 0.9 n_0) Gaussian bunch impacting a homogeneous plasma of density `5.2e24`.

Boths simulations started from the same checkpoint at time step 12461. Radiation computation starts at time step 17000 while the bunch is still in vacuo.
It impacts the background plasma around time step 19961.

In the following images the reference spectrum is plotted on the left, the new spectrum is shown on the right and their difference in between.
The top color bar pertains to the intensity spectra and the bottom to their difference (symm-log). The black arrow in the rightmost figure signifies the detector/observer direction and the magenta arrow the direction of propagation of the bunch).
![CMP_CumulativeIntensity_detectorNr_0](https://user-images.githubusercontent.com/50339174/107638938-8943f000-6c70-11eb-94c4-cf1364fba7eb.png)

![CMP_CumulativeIntensity_detectorNr_3](https://user-images.githubusercontent.com/50339174/107639110-bf816f80-6c70-11eb-94d0-f59462ae80f3.png)

As can be seen the difference appears significant for the detector being co-linear with the bunch propagation. Differences decay as soon as observer and bunch propagation are not collinear.

If one considers the intensity integrated over the frequencies for each time step the following are obtained for the observer directions shown above:
![CMP_IntegratedIntensityVsTimeStep_0](https://user-images.githubusercontent.com/50339174/107639439-24d56080-6c71-11eb-9b5f-dbc19b1d974d.png)
![CMP_IntegratedIntensityVsTimeStep_3](https://user-images.githubusercontent.com/50339174/107639448-27d05100-6c71-11eb-9ca9-473967d938e6.png)

**Note** In this case the vertical axis is linear.