@psychocoderHPC 2021-02-11
@Anton-Le I butified your nice PR description, you can render latex in github: https://gist.github.com/a-rodin/fef3f543412d6e1ec5b6cf55bf197d7b