@psychocoderHPC 2021-02-11
> typename PICToSplash<float_64>::type radSplashType

Please link the code you mentioned `typename PICToSplash<float_64>::type radSplashType`