@psychocoderHPC 2021-02-11
> @psychocoderHPC Is it possible to parametrize the data type (here `float_64`) of the `typename PICToSplash<float_64>::type radSplashType;` in `Radiation.hpp` to always be determined by the type of of the data stored in `detectorPositions` and `detectorFrequencies` ? Ideally without templating the entire class.
> Currently I just introduced 3 different `typename` statements, which does not benefit maintainability.

I do not understand the problem `typename PICToSplash<MyRadType>::type radSplashType;` will be possible. You need to have `MyRadType` somewhere in a header if you need to include it within many files.