@sbastrakov 2021-02-17
I think the proposed changes in this PR seem very reasonable and do not require changing that many places in the code, so nothing against it. The following is a bit theoretical, and I point out its limitations below, but just to share and float an idea. We may have a similar pattern in SAXS btw.

I think it is possible to use only single precision for summation, if the accumulator variables are updated compensated summation like `2sum`. I now found my own [very old materians on that in Russian](https://software.intel.com/sites/default/files/m/d/4/1/d/8/LW_CA_Summation_doc.pdf).
Basically, the idea is very simple, to copy relevant code fragments from that document:
```cpp
// 2Sum(a, b): a+ b= s+ t
// https://en.wikipedia.org/wiki/2Sum
void twoSum(float a, float b, float& s, float& t){
    s = a + b;
    float a_ = s -b;
    float b_ = s -a;
    float da = a -a_;
    float db = b -b_;
    t = da + db;
}

float sumFloatCompensated(const float* x, int n){
    float sum = x[0];
    float c = 0;
    for(int i = 1; i < n; ++i) {
        float y = x[i] + c;
        twoSum(sum, y, sum, c);
    }
    return sum;
}
```
Accuracy-wise it should be mostly same than a "stupid" summation with `double`, as long as each separate value is well-represented as `float`. For absurdly large number of terms will actually be better as there is no running accumulated error in this scheme (as it's compensated each time), but I don't think we are nearly close to that. So practically I think it would be same as adding everything up in `double`. Performance-wise it is worse than just using `double`, also worse register-wise and a couple more flops (the latter very negligible in the context of this plugin). But it may save GPU device memory for the results of summations as these are only `float`. There are two obstacles to use it right now in the radiation plugin:

- The structure of computations there is such that each thread iterates first through all supercells, and then through its own frequencies. So with this loop order there needs to be an additional accumulator for each frequency handled by a thread, and that is not a fixed amount. If the loops were inversed: first frequencies, then supercells that would not be a problem, as then only a single register `float` would suffice
- It needs to be checked how it works on GPUs and with optimization flags.